+++ lib/govuk_publishing_components/presenters/taxonomy_navigation.rb
@@ -29,22 +29,8 @@ module GovukPublishingComponents
         end
       end
 
-      def there_are_related_item_overrides?
-        # TODO: We should check for any external links when we have "new"
-        # external links being curated in Content Tagger
-        @content_item.curated_taxonomy_sidebar_links.any?
-      end
-
       def related_items
-        related_items_factory.new(@content_item).related_items
-      end
-
-      def related_items_factory
-        if there_are_related_item_overrides?
-          CuratedTaxonomySidebarLinks
-        else
-          RummagerTaxonomySidebarLinks
-        end
+        CuratedTaxonomySidebarLinks.new(@content_item).related_items
       end
 
       def collections